+++ llvm/docs/ReleaseNotes.md
@@ -220,8 +220,7 @@ Changes to LLDB
 * Threads are listed in incrmental order by pid then by tid.
 * Unread kernel messages saved in msgbufp are now printed when lldb starts. This information is printed only
   when lldb is in the interactive mode (i.e. not in batch mode).
-* Writing to core is now supported. For security, `plugin.process.freebsd-kernel-core.read-only` must be set
-  to `false` to enable this feature.
+* Memory writes are currently disabled. You can enable them with `set plugin.process.freebsd-kernel-core.read-only false`.
 
 ### Linux